# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The REST API a Quark device serves to its Flutter clients. Every endpoint except\n/auth/setup, /auth/login, /auth/recover, /auth/request-account and /auth/status needs a\nsession token. Sign in with POST /auth/login, then click Authorize and enter the word\nBearer, a space, and the token.",
        "title": "Quark API",
        "contact": {},
        "version": "v0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/videos/metadata": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns duration, resolution, codec, bitrate, framerate, rotation, and the caller's own favorite state and album membership for the specified video.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get metadata for a single video file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device serial",
                        "name": "serial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Relative path to the video file",
                        "name": "relPath",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v0_videos.VideoMetadataJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented — ffprobe not available",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "v0_videos.VideoMetadataJSON": {
            "type": "object",
            "properties": {
                "albums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v0_videos.albumRefJSON"
                    }
                },
                "audioCodec": {
                    "type": "string"
                },
                "bitrate": {
                    "description": "bits/s",
                    "type": "integer"
                },
                "duration": {
                    "description": "seconds",
                    "type": "number"
                },
                "fileName": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "framerate": {
                    "type": "number"
                },
                "height": {
                    "type": "integer"
                },
                "isFavorite": {
                    "type": "boolean"
                },
                "mtime": {
                    "type": "integer"
                },
                "rotation": {
                    "description": "degrees",
                    "type": "integer"
                },
                "videoCodec": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "v0_videos.albumRefJSON": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "A session token from POST /auth/login, entered as the word Bearer, a space, and the token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}